Move update log into WMT as a tab; drop standalone page
Per request, the update log now lives inside the WMT tool as an
"Update-Log" tab (gated to matmaxx.org via isUpdateLogDomain(), and the
tab row is already hidden on the opponents' mirror) instead of a top-level
nav-card on the Home page.

- New reusable component UpdateLogList renders the scheduler-run rows
  (incl. next-run reason); used by the WMT tab.
- Removed the standalone /update-log route, the UpdateLog page, and the
  Home nav-card.
- wmt v3.25 → v3.26.
